# cuchucambiazo/src/main/resources/openapi/api-media.yml
openapi: 3.0.0
info:
  title: API Business Layer
  version: "0.1"
paths:

  /media:

    post:
      summary: Get Medias
      tags:
        - media
      operationId: getMedias

      requestBody:
        description: Save likes request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/getMediaRequest"
      responses:
        '201':
          description: Add user confirmation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/getMediaResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '404':
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'

  /media/save:

    post:
      summary: Save medias
      tags:
        - media
      operationId: saveMedia

      requestBody:
        description: Save medias request
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MediaRequest"
      responses:
        '200':
          description: Add user confirmation
          content:
            application/json:
              schema:
                $ref: 'api-user.yml#/components/schemas/GeneralBusinessResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '404':
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'

  /media/delete:

    post:
      summary: Remove medias
      tags:
        - media
      operationId: removeMedia

      requestBody:
        description: Save medias request
        content:
          text/plain:
            schema:
              type: integer
      responses:
        '201':
          description: Remove user confirmation
          content:
            application/json:
              schema:
                $ref: 'api-user.yml#/components/schemas/GeneralBusinessResponse'
        '400':
          description: Bad Request.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '404':
          description: Not Found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'
        '500':
          description: Internal Server Error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralBusinessError'



components:
  schemas:



    getMediaRequest:
      type: object
      description: ''
      required:
        - isHome
        - userId
      properties:
        isHome:
          type: boolean
        email:
          type: string

    getMediaResponse:
      description: ''
      type: object
      properties:
        media_list:
          $ref: '#/components/schemas/MediaList'
      required:
        - media_list

    MediaList:
      type: array
      items:
        $ref: '#/components/schemas/Media'

    MediaRequest:
      description: ''
      type: object
      properties:
        email:
          type: string
        media:
          $ref: '#/components/schemas/Media'
      required:
        - Media
        - email

    Media:
      type: object
      description: ''
      required:
        - photoList
        - state
        - description
        - measures
      properties:
        mediaId:
          type: integer
        userId:
          type: integer
        title:
          type: string
        color:
          type: string
        brand:
          type: string
        photoList:
          type: array
          items:
            $ref: '#/components/schemas/Photo'
        state:
          type: string
        description:
          type: string
        measures:
          type: string
        tag:
          type: string
        updateDate:
          type: string
        dischargeDate:
          type: string

    Photo:
      type: object
      description: ''
      properties:
        photoId:
          type: integer
        mediaId:
          type: integer
        url:
          type: string
          format: base64

    GeneralBusinessError:
      type: object
      description: 'Code that represents the error that took place'
      properties:
        code:
          type: string
        message:
          type: string
          description: Human readable description of the error